[Break Change] Optimize performance
The algorithm is reimplemented, using the tr() function instead of loop,
which is about 30 times faster.

------------------------------------------------------------

This plugin is created in my early stage, at that time I was completely
a newbie in vim, so I directly placed the function in `/plugin` directory.

It's not a good idea, because by this way the function will always be
loaded, which may slow down the startup of vim. In this commit, I placed
the function in `/autoload` directory, so it will be loaded on demand.

But this means you'll need to modify your vimrc. All you need to to is
simply replace `Artify()` with `artify#convert()`, like this:

```vim
Artify('Introduction', 'italic')  " old
artify#convert('Introduction', 'italic')  " new
```

------------------------------------------------------------

Btw, for neovim users, consider switch to this lua implementation which is
much more faster: https://github.com/delphinus/artify.nvim